third_party/WebKit/LayoutTests/fast/dom/SelectorAPI/dumpNodeList-2-expected.txt

line 4

line 5

PASS testQuerySelectorAll(document, ".target + div") is "[object HTMLDivElement], [object HTMLDivElement], length: 2"
PASS testQuerySelectorAll(document.getElementById("test"), ".target + div") is "[object HTMLDivElement], [object HTMLDivElement], length: 2"
PASS testQuerySelectorAll(document.getElementById("test2"), ".target + div") is "length: 0"
PASS testQuerySelectorAll(document.getElementById("test3"), ".target + div") is "length: 0"
PASS successfullyParsed is true

TEST COMPLETE
